AudioCore: Buffer management

+++ src/audio_core/audio_core.h
@@ -10,9 +10,9 @@ class VMManager;
 
 namespace AudioCore {
 
-constexpr int num_sources = 24;
-constexpr int samples_per_frame = 160;     ///< Samples per audio frame at native sample rate
-constexpr int native_sample_rate = 32728;  ///< 32kHz
+constexpr size_t num_sources = 24;
+constexpr size_t samples_per_frame = 160; ///< Samples per audio frame at native sample rate
+constexpr unsigned native_sample_rate = 32728; ///< 32kHz
 
 /// Initialise Audio Core
 void Init();
